improve default compare/official scripts

+++ sources/official/compare.sh
@@ -2,8 +2,10 @@
 
 cd $(dirname $0)
 
+[ -s official.html ] || exit
+
 bundle exec ruby scraper.rb | ifne tee scraped.csv
 wd sparql -f csv wikidata.js | sed -e 's/T00:00:00Z//g' -e 's#http://www.wikidata.org/entity/##g' | qsv dedup -s psid | ifne tee wikidata.csv
 bundle exec ruby diff.rb | tee diff.csv
 
-cd -
+cd ~-
